Add more validation for country pages, fix error handling, decouple components
