+++ pages/_error.js
@@ -1,8 +1,11 @@
 import * as Sentry from '@sentry/nextjs'
+import NavBar from 'components/NavBar'
 import NextErrorComponent from 'next/error'
-
 const ErrorPage = (props) => (
-  <NextErrorComponent statusCode={props.statusCode} title={props.error} />
+  <>
+    <NavBar />
+    <NextErrorComponent statusCode={props.statusCode} title={props.error} />
+  </>
 )
 
 ErrorPage.getInitialProps = async (contextData) => {
